requirements.txt
altair
asttokens
attrs
blinker
Bottleneck
cachetools
certifi
charset-normalizer
click
comm
datetime
debugpy
decorator
et_xmlfile
executing
gitdb
GitPython
idna
ipykernel
ipython
ipython_pygments_lexers
jedi
Jinja2
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
MarkupSafe
matplotlib-inline
mkl-service
mkl_fft
mkl_random
narwhals
nest-asyncio
numexpr
numpy
oauthlib
openpyxl
typing
packaging
pandas
parso
pexpect
pillow
platformdirs
prompt_toolkit
protobuf
psutil
ptyprocess
pure_eval
pyarrow
pydeck
Pygments
python-dateutil
python-dotenv
pytz
pyzmq
referencing
requests
requests-oauthlib
rpds-py
setuptools
six
smmap
stack-data
streamlit==1.34.0
tenacity
toml
tornado
traitlets
typing_extensions
tzdata
urllib3
watchdog
wcwidth
wheel
XlsxWriter
